third_party/WebKit/LayoutTests/http/tests/inspector/network/network-initiator-expected.txt

CONSOLE MESSAGE: line 14: Done.
Tests resources initiator for images initiated by IMG tag, static CSS, CSS class added from JavaScript and XHR.

Bug 65105  
http://127.0.0.1:8000/inspector/network/resources/initiator.css: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 2
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=100: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 4
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=400: script
    loadData http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 12
http://127.0.0.1:8000/inspector/network/resources/style.css: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 7
http://127.0.0.1:8000/inspector/network/resources/empty.html: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 16
